started to fix plants

+++ src/main/java/twilightforest/entity/EntityTFTowerGhast.java
@@ -264,15 +264,15 @@ public class EntityTFTowerGhast extends EntityGhast {
 	}
 
 	public int getAttackStatus() {
-		return dataManager.get(ATTACK_STATUS);
+		return (int)dataManager.get(ATTACK_STATUS);
 	}
 
 	public int getAttackTimer() {
-		return dataManager.get(ATTACK_TIMER);
+		return (int)dataManager.get(ATTACK_TIMER);
 	}
 
 	public int getPrevAttackTimer() {
-		return dataManager.get(ATTACK_PREVTIMER);
+		return (int)dataManager.get(ATTACK_PREVTIMER);
 	}
 
 	protected boolean shouldAttack(EntityLivingBase living) {
